backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==3.2.2
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.118.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.3
openpyxl==3.1.5
orjson==3.11.3
pandas==2.3.3
passlib==1.7.4
pycparser==2.23
pydantic==2.12.0
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
sentry-sdk==2.40.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
razorpay==1.4.1
aiosmtplib==2.0.2
redis==5.2.1
fpdf2>=2.8.0
setuptools<70
